branches-api/corte-caja-api/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0

python-dotenv==1.0.1
requests==2.31.0

SQLAlchemy==2.0.23
psycopg2-binary==2.9.9

python-jose==3.3.0
python-multipart==0.0.9